+++ app/Http/Controllers/Admin/AdminMarimeCompanieController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CompanieSize;
+use App\Models\Company;
 
 class AdminMarimeCompanieController extends Controller
 {
@@ -53,6 +54,11 @@ class AdminMarimeCompanieController extends Controller
     }
 
     public function stergere($id){
+        $verificare = Company::where('companie_size_id',$id)->count();
+        if($verificare>0)
+        {
+            return redirect()->back()->with('error','Aceasta locatie este folosita deci nu poate fi stearsa!');
+        }
         
         $marime_companie = CompanieSize::where('id',$id)->delete();
